Set up messente namespace (#1)
