+++ .gitignore
@@ -5,3 +5,4 @@ build/
 *.egg
 *.egg-info/
 *.log
+.idea
