+++ messente/api/sms/messente.py
@@ -16,13 +16,21 @@
 
 from __future__ import absolute_import, division, print_function
 
-from . import error
-from .response import Response
+from messente.api.sms import api
 
-# api modules
-from . import sms
-from . import credit
-from . import delivery
-from . import pricing
-from . import number_verification
-from . import verification_widget
+
+class Messente(object):
+    def __init__(self, **kwargs):
+        if kwargs.get("ini_path", ""):
+            api.config.load(kwargs.pop("ini_path"))
+        # modules
+        self.sms = api.sms.SmsAPI(**kwargs)
+        self.credit = api.credit.CreditAPI(**kwargs)
+        self.delivery = api.delivery.DeliveryAPI(**kwargs)
+        self.pricing = api.pricing.PricingAPI(**kwargs)
+        self.number_verification = (
+            api.number_verification.NumberVerificationAPI(**kwargs)
+        )
+        self.verification_widget = (
+            api.verification_widget.VerificationWidgetAPI(**kwargs)
+        )
