+++ messente/api/sms/test/test_number_verification.py
@@ -2,12 +2,9 @@
 
 import responses
 
-from test import utils
-
-from messente.api import number_verification
-from messente.api import Response
-from messente.api.error import ConfigurationError
-
+from messente.api.sms.api import Response
+from messente.api.sms.api import number_verification
+from messente.api.sms.test import utils
 
 api = number_verification.NumberVerificationAPI(
     user="test",
