+++ messente/api/sms/test/test_number_verification_validate_pin.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import unittest
-import time
-from messente.api import number_verification
+
+from messente.api.sms.api import number_verification
 
 
 class TestValidate(unittest.TestCase):
