+++ messente/api/sms/test/test_pricing.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
 
-import responses
 import json
 import tempfile
 
-from test import utils
+import responses
 
-from messente.api import pricing
-from messente.api import Response
-
+from messente.api.sms.api import Response
+from messente.api.sms.api import pricing
+from messente.api.sms.test import utils
 
 fake_response = json.dumps({
     "country": "EE",
