+++ messente/api/sms/test/test_sms_send.py
@@ -2,12 +2,10 @@
 
 import responses
 
-from test import utils
-
-from messente.api import sms
-from messente.api import Response
-from messente.api.error import ConfigurationError
-from messente.api.error import InvalidMessageError
+from messente.api.sms.api import Response
+from messente.api.sms.api import sms
+from messente.api.sms.api.error import InvalidMessageError
+from messente.api.sms.test import utils
 
 
 def mk_sms_data(data=None):
