+++ messente/api/sms/test/utils.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import responses
 import re
-
 
 TEST_BASE_URL = "https://example.com"
 TEST_URL = "%s/api" % TEST_BASE_URL
@@ -14,4 +12,5 @@ TEST_DLR_ANY_URL = re.compile(TEST_DLR_URL + "/.*")
 def mock_response(http_code, response_text, **kwargs):
     def request_callback(request):
         return (http_code, kwargs.get("headers", {}), str(response_text))
+
     return request_callback
